+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change Log
 
 * Fixed issue of highwire elements storing escaped `content` as `bytes`. [#50](https://github.com/unt-libraries/pyuntl/pull/50)
 * Replaced deprecated `cgi.escape` with `html.escape`. [#50](https://github.com/unt-libraries/pyuntl/pull/50)
+* Moved the logic of FormGenerator.get_vocabularies to a new function, which is now called by the deprecated
+  FormGenerator.get_vocabularies method and by retrieve_vocab.
+* Modified new get_vocabularies function to use retries when retrieving the vocabularies, as well as caching them
+  locally.
 
 2.0.0
 -----